pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "inbox-copilot"
version = "0.1.0"
description = "Local-first Gmail inbox automation with rule-based and AI-assisted classification"
requires-python = ">=3.10"

dependencies = [
    "openai>=1.50.0,<2.0.0",
    "python-dotenv",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
]

[tool.setuptools.packages.find]
where = ["src"]
